requirements.txt
django
djangorestframework
djangorestframework-simplejwt[crypto]
